+++ RELEASE-NOTES.md
@@ -3,7 +3,7 @@
 
 ## 1.3.0
 
-* refactor: http请求失败重试间隔优化。
+* refactor: http请求失败重试间隔优化。TokenExpired 和 ShareLinkTokenInvalid 不再无限重试，改成递减。
 * refactor: createFolders 去掉返回字符串逻辑，统一抛出异常。
 
 ## 1.2.0